refacto tetris game hook into smaller parts
